Add check fragment drop ability
This adds a config option to drop fragmented packets. The check itself
is simple and looks if MF (more frags) is 1 or fragment offset is > 0
and then drops the packets.

I haven't done performance testing of this code but I suspect it is
ridiculously fast so it's a great way to drop lots of traffic fast if
you aren't expecting fragments!

Fixes #8