Add basic packet validity checks
We check that the IP version, as read from the IP header, is 4. Since we
only support IPv4 right now, we simply throw away packets with other
values since that means they are invalid. Note that before this we
already check the ethertype and forward packets that have an ethertype
indicating something else than IPv4, so this check really only makes
sure that the IP version as read from IP header is same as advertised in
the ethertype.

The second check looks at the packet length and makes sure it is what
it's supposed to be. It only does the comparison for packets longer than
60 bytes since shorter packets are padded and we have no real way of
validity checking that.